LayoutTests/fast/profiler/calling-the-function-that-started-the-profiler-from-another-scope-expected.txt

CONSOLE MESSAGE: line 59: Calling the same function where the profile started from another function
CONSOLE MESSAGE: line 75: Thread_1  0
CONSOLE MESSAGE: line 75:  indirection calling-the-function-that-started-the-profiler-from-another-scope.html 16
CONSOLE MESSAGE: line 75:   functionWichStartsAndStopsTheProfiler calling-the-function-that-started-the-profiler-from-another-scope.html 22
CONSOLE MESSAGE: line 75:  functionWichStartsAndStopsTheProfiler calling-the-function-that-started-the-profiler-from-another-scope.html 22
CONSOLE MESSAGE: line 75:  endTest profiler-test-JS-resources.js 1
CONSOLE MESSAGE: line 75:  (idle)  0
This page's JavaScript starts profiling from within a function which is then called later from another context. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profiler you should see a call to indirection() and functionWichStartsAndStopsTheProfiler()
